move table creation/insertion
